requirements.txt
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
aliyun-python-sdk-alidns==2.6.19
aliyun-python-sdk-core==2.13.29
huaweicloudsdkcore==3.1.5
huaweicloudsdkdns==3.1.5
tencentcloud-sdk-python==3.0.806